requirements.txt
# Django Core
Django==5.2.7

# ASGI Server & WebSockets
daphne==4.1.2
channels==4.2.0
channels-redis==4.2.1

# Database
psycopg2-binary==2.9.10

# REST Framework
djangorestframework==3.15.2
django-filter==24.3

# CORS Headers
django-cors-headers==4.6.0

# Translation
django-modeltranslation==0.19.12

# Environment Variables
python-dotenv==1.0.1

# Production Server
uWSGI==2.0.28

# Google Gemini AI
google-generativeai==0.8.3

# WebSocket Support
websockets==14.1

# Redis for Channels Layer
redis==5.2.1

# Additional dependencies
Pillow==11.0.0  # For image handling

# AWS S3 Storage
boto3==1.35.94
django-storages==1.14.4
